Fix incorrect session initialization for YandexDriving, YandexBicycle, YandexSuggest, YandexSearch. Before we were returning session before it was fully initialized on native side, which in turn brought errors like #348
