+++ ios/Classes/YandexBicycle.swift
@@ -6,7 +6,6 @@ public class YandexBicycle: NSObject, FlutterPlugin {
   private let methodChannel: FlutterMethodChannel!
   private let pluginRegistrar: FlutterPluginRegistrar!
   private let bicycleRouter: YMKBicycleRouter!
-  private var bicycleSessions: [Int: YandexBicycleSession] = [:]
 
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(
@@ -31,33 +30,19 @@ public class YandexBicycle: NSObject, FlutterPlugin {
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
-    case "requestRoutes":
-      requestRoutes(call, result)
+    case "initSession":
+      initSession(call)
+      result(nil)
       break
     default:
       result(FlutterMethodNotImplemented)
     }
   }
 
-  private func requestRoutes(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
+  public func initSession(_ call: FlutterMethodCall) {
     let params = call.arguments as! [String: Any]
-    let sessionId = params["sessionId"] as! Int
-    let requestPoints = (params["points"] as! [[String: Any]]).map {
-      (pointParams) -> YMKRequestPoint in Utils.requestPointFromJson(pointParams)
-    }
-    let session = bicycleRouter.requestRoutes(
-      with: requestPoints,
-      type: YMKBicycleVehicleType(rawValue: (params["bicycleVehicleType"] as! NSNumber).uintValue)!,
-      routeListener: {(bicycleResponse: [YMKBicycleRoute]?, error: Error?) -> Void in
-        self.bicycleSessions[sessionId]?.handleResponse(bicycleResponse: bicycleResponse, error: error, result: result)
-      }
-    )
+    let id  = params["id"] as! Int
 
-    bicycleSessions[sessionId] = YandexBicycleSession(
-      id: sessionId,
-      session: session,
-      registrar: pluginRegistrar,
-      onClose: { (id) in self.bicycleSessions.removeValue(forKey: id) }
-    )
+    YandexBicycleSession.initSession(id: id, registrar: pluginRegistrar, bicycleRouter: bicycleRouter)
   }
 }
