+++ ios/Classes/YandexDriving.swift
@@ -6,7 +6,6 @@ public class YandexDriving: NSObject, FlutterPlugin {
   private let methodChannel: FlutterMethodChannel!
   private let pluginRegistrar: FlutterPluginRegistrar!
   private let drivingRouter: YMKDrivingRouter!
-  private var drivingSessions: [Int: YandexDrivingSession] = [:]
 
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(
@@ -31,34 +30,19 @@ public class YandexDriving: NSObject, FlutterPlugin {
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
-    case "requestRoutes":
-      requestRoutes(call, result)
+    case "initSession":
+      initSession(call)
+      result(nil)
       break
     default:
       result(FlutterMethodNotImplemented)
     }
   }
 
-  private func requestRoutes(_ call: FlutterMethodCall, _ result: @escaping FlutterResult) {
+  public func initSession(_ call: FlutterMethodCall) {
     let params = call.arguments as! [String: Any]
-    let sessionId = params["sessionId"] as! Int
-    let requestPoints = (params["points"] as! [[String: Any]]).map {
-      (pointParams) -> YMKRequestPoint in Utils.requestPointFromJson(pointParams)
-    }
-    let session = drivingRouter.requestRoutes(
-      with: requestPoints,
-      drivingOptions: Utils.drivingOptionsFromJson(params["drivingOptions"] as! [String: Any]),
-      vehicleOptions: YMKDrivingVehicleOptions(),
-      routeHandler: {(drivingResponse: [YMKDrivingRoute]?, error: Error?) -> Void in
-        self.drivingSessions[sessionId]?.handleResponse(drivingResponse: drivingResponse, error: error, result: result)
-      }
-    )
+    let id  = params["id"] as! Int
 
-    drivingSessions[sessionId] = YandexDrivingSession(
-      id: sessionId,
-      session: session,
-      registrar: pluginRegistrar,
-      onClose: { (id) in self.drivingSessions.removeValue(forKey: id) }
-    )
+    YandexDrivingSession.initSession(id: id, registrar: pluginRegistrar, drivingRouter: drivingRouter)
   }
 }
