+++ ios/Classes/YandexSuggest.swift
@@ -3,12 +3,10 @@ import Flutter
 import UIKit
 import YandexMapsMobile
 
-
 public class YandexSuggest: NSObject, FlutterPlugin {
   private let pluginRegistrar: FlutterPluginRegistrar!
   private let methodChannel: FlutterMethodChannel!
   private let searchManager: YMKSearchManager!
-  private var suggestSessions: [Int: YandexSuggestSession] = [:]
 
   public static func register(with registrar: FlutterPluginRegistrar) {
     let channel = FlutterMethodChannel(
@@ -33,32 +31,18 @@ public class YandexSuggest: NSObject, FlutterPlugin {
 
   public func handle(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
     switch call.method {
-    case "getSuggestions":
-      getSuggestions(call, result: result)
+    case "initSession":
+      initSession(call)
+      result(nil)
     default:
       result(FlutterMethodNotImplemented)
     }
   }
 
-  public func getSuggestions(_ call: FlutterMethodCall, result: @escaping FlutterResult) {
+  public func initSession(_ call: FlutterMethodCall) {
     let params = call.arguments as! [String: Any]
-    let sessionId = (params["sessionId"] as! NSNumber).intValue
-    let session = self.searchManager!.createSuggestSession()
+    let id  = params["id"] as! Int
 
-    session.suggest(
-      withText: params["text"] as! String,
-      window: Utils.boundingBoxFromJson(params["boundingBox"] as! [String: Any]),
-      suggestOptions: Utils.suggestOptionsFromJson(params["suggestOptions"] as! [String: Any]),
-      responseHandler: {(suggestResponse: [YMKSuggestItem]?, error: Error?) -> Void in
-        self.suggestSessions[sessionId]?.handleResponse(suggestResponse: suggestResponse, error: error, result: result)
-      }
-    )
-
-    suggestSessions[sessionId] = YandexSuggestSession(
-      id: sessionId,
-      session: session,
-      registrar: pluginRegistrar,
-      onClose: { (id) in self.suggestSessions.removeValue(forKey: id) }
-    )
+    YandexSuggestSession.initSession(id: id, registrar: pluginRegistrar, searchManager: searchManager)
   }
 }
