+++ lib/src/types/bicycle/bicycle_session.dart
@@ -6,52 +6,37 @@ class BicycleSession {
 
   /// Unique session identifier
   final int id;
-  bool _isClosed = false;
-
-  /// Has the current session been closed
-  bool get isClosed => _isClosed;
 
   BicycleSession._({required this.id}) :
     _methodChannel = MethodChannel(_methodChannelName + id.toString());
 
   /// Retries current session
-  ///
-  /// After [BicycleSession.close] has been called, all subsequent calls will return a [BicycleSessionException]
   Future<void> retry() async {
-    if (_isClosed) {
-      throw BicycleSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('retry');
   }
 
   /// Cancels current session
-  ///
-  /// After [BicycleSession.close] has been called, all subsequent calls will return a [BicycleSessionException]
   Future<void> cancel() async {
-    if (_isClosed) {
-      throw BicycleSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('cancel');
   }
 
   /// Closes current session
-  ///
-  /// After first call, all subsequent calls will return a [BicycleSessionException]
   Future<void> close() async {
-    if (_isClosed) {
-      throw BicycleSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('close');
-
-    _isClosed = true;
   }
-}
 
-class BicycleSessionException extends SessionException {
-  BicycleSessionException._(String message) : super._(message);
+  Future<BicycleSessionResult> _requestRoutes({
+    required List<RequestPoint> points,
+    required BicycleVehicleType bicycleVehicleType
+  }) async {
+    final params = <String, dynamic>{
+      'bicycleVehicleType': bicycleVehicleType.index,
+      'points': points.map((RequestPoint requestPoint) => requestPoint.toJson()).toList(),
+    };
+    final result = await _methodChannel.invokeMethod('requestRoutes', params);
+
+    return BicycleSessionResult._fromJson(result);
+  }
 }
 
 /// Result of a request to build routes
@@ -71,19 +56,4 @@ class BicycleSessionResult {
       json['error']
     );
   }
-}
-
-/// Object containing the result of a route building request and
-/// a [session] object for further working with newly made request
-class BicycleResultWithSession {
-  /// Created session
-  BicycleSession session;
-
-  /// Request result
-  Future<BicycleSessionResult> result;
-
-  BicycleResultWithSession._({
-    required this.session,
-    required this.result
-  });
 }
