+++ lib/src/types/driving/driving_session.dart
@@ -6,52 +6,38 @@ class DrivingSession {
 
   /// Unique session identifier
   final int id;
-  bool _isClosed = false;
-
-  /// Has the current session been closed
-  bool get isClosed => _isClosed;
 
   DrivingSession._({required this.id}) :
     _methodChannel = MethodChannel(_methodChannelName + id.toString());
 
   /// Retries current session
-  ///
-  /// After [DrivingSession.close] has been called, all subsequent calls will return a [DrivingSessionException]
   Future<void> retry() async {
-    if (_isClosed) {
-      throw DrivingSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('retry');
   }
 
   /// Cancels current session
-  ///
-  /// After [DrivingSession.close] has been called, all subsequent calls will return a [DrivingSessionException]
   Future<void> cancel() async {
-    if (_isClosed) {
-      throw DrivingSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('cancel');
   }
 
   /// Closes current session
-  ///
-  /// After first call, all subsequent calls will return a [DrivingSessionException]
   Future<void> close() async {
-    if (_isClosed) {
-      throw DrivingSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('close');
-
-    _isClosed = true;
   }
-}
 
-class DrivingSessionException extends SessionException {
-  DrivingSessionException._(String message) : super._(message);
+  Future<DrivingSessionResult> _requestRoutes({
+    required List<RequestPoint> points,
+    required DrivingOptions drivingOptions
+  }) async {
+    final params = <String, dynamic>{
+      'points': points.map((RequestPoint requestPoint) => requestPoint.toJson()).toList(),
+      'drivingOptions': drivingOptions.toJson()
+    };
+
+    final result = await _methodChannel.invokeMethod('requestRoutes', params);
+
+    return DrivingSessionResult._fromJson(result);
+  }
 }
 
 /// Result of a request to build routes
@@ -71,19 +57,4 @@ class DrivingSessionResult {
       json['error']
     );
   }
-}
-
-/// Object containing the result of a route building request and
-/// a [session] object for further working with newly made request
-class DrivingResultWithSession {
-  /// Created session
-  DrivingSession session;
-
-  /// Request result
-  Future<DrivingSessionResult> result;
-
-  DrivingResultWithSession._({
-    required this.session,
-    required this.result
-  });
 }
