+++ lib/src/types/suggest/suggest_session.dart
@@ -7,41 +7,35 @@ class SuggestSession {
 
   /// Unique session identifier
   final int id;
-  bool _isClosed = false;
-
-  /// Has the current session been closed
-  bool get isClosed => _isClosed;
 
   SuggestSession._({required this.id}) :
     _methodChannel = MethodChannel(_methodChannelName + id.toString());
 
   /// Resets current session
-  ///
-  /// After [SuggestSession.close] has been called, all subsequent calls will return a [SuggestSessionException]
   Future<void> reset() async {
-    if (_isClosed) {
-      throw SuggestSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('reset');
   }
 
   /// Closes current session
-  ///
-  /// After first call, all subsequent calls will return a [SuggestSessionException]
   Future<void> close() async {
-    if (_isClosed) {
-      throw SuggestSessionException._('Session is closed');
-    }
-
     await _methodChannel.invokeMethod<void>('close');
-
-    _isClosed = true;
   }
-}
 
-class SuggestSessionException extends SessionException {
-  SuggestSessionException._(String message) : super._(message);
+  Future<SuggestSessionResult> _getSuggestions({
+    required String text,
+    required BoundingBox boundingBox,
+    required SuggestOptions suggestOptions
+  }) async {
+    final params = <String, dynamic>{
+      'text': text,
+      'boundingBox': boundingBox.toJson(),
+      'suggestOptions': suggestOptions.toJson(),
+    };
+
+    final result = await _methodChannel.invokeMethod('getSuggestions', params);
+
+    return SuggestSessionResult._fromJson(result);
+  }
 }
 
 /// Result of a suggest request
@@ -62,19 +56,4 @@ class SuggestSessionResult {
       json['error']
     );
   }
-}
-
-/// Object containing the result of a suggest request and
-/// a [session] object for further working with newly made request
-class SuggestResultWithSession {
-  /// Created session
-  SuggestSession session;
-
-  /// Request result
-  Future<SuggestSessionResult> result;
-
-  SuggestResultWithSession._({
-    required this.session,
-    required this.result
-  });
 }
