+++ lib/src/yandex_search.dart
@@ -5,49 +5,36 @@ class YandexSearch {
   static const String _channelName = 'yandex_mapkit/yandex_search';
   static const MethodChannel _channel = MethodChannel(_channelName);
 
-  static int _nextSessionId = 0;
+  static int _nextId = 0;
 
   /// Search request for searching a user query near given geometry.
-  static SearchResultWithSession searchByText({
+  static Future<(SearchSession, Future<SearchSessionResult>)> searchByText({
     required String searchText,
     required Geometry geometry,
     required SearchOptions searchOptions
-  }) {
-    final params = <String, dynamic>{
-      'sessionId': _nextSessionId++,
-      'searchText': searchText,
-      'geometry': geometry.toJson(),
-      'searchOptions': searchOptions.toJson(),
-    };
-    final result = _channel
-      .invokeMethod('searchByText', params)
-      .then((result) => SearchSessionResult._fromJson(result));
+  }) async {
+    final session = await _initSession();
 
-    return SearchResultWithSession._(
-      session: SearchSession._(id: params['sessionId']),
-      result: result,
-    );
+    return (session, session._searchByText(searchText: searchText, geometry: geometry, searchOptions: searchOptions));
   }
 
   /// Reverse search request (to search objects at the given coordinates)
-  static SearchResultWithSession searchByPoint({
+  static Future<(SearchSession, Future<SearchSessionResult>)> searchByPoint({
     required Point point,
     int? zoom,
     required SearchOptions searchOptions
-  }) {
-    final params = <String, dynamic>{
-      'sessionId': _nextSessionId++,
-      'point': point.toJson(),
-      'zoom': zoom,
-      'searchOptions': searchOptions.toJson(),
-    };
-    final result = _channel
-      .invokeMethod('searchByPoint', params)
-      .then((result) => SearchSessionResult._fromJson(result));
+  }) async {
+    final session = await _initSession();
 
-    return SearchResultWithSession._(
-      session: SearchSession._(id: params['sessionId']),
-      result: result,
-    );
+    return (session, session._searchByPoint(point: point, searchOptions: searchOptions));
+  }
+
+  /// Initialize session on native side for further use
+  static Future<SearchSession> _initSession() async {
+    final id = _nextId++;
+
+    await _channel.invokeMethod('initSession', { 'id': id });
+
+    return SearchSession._(id: id);
   }
 }
