+++ lib/src/yandex_suggest.dart
@@ -5,27 +5,25 @@ class YandexSuggest {
   static const String _channelName = 'yandex_mapkit/yandex_suggest';
   static const MethodChannel _channel = MethodChannel(_channelName);
 
-  static int _nextSessionId = 0;
+  static int _nextId = 0;
 
   /// Get suggestions for text
-  static SuggestResultWithSession getSuggestions({
+  static Future<(SuggestSession, Future<SuggestSessionResult>)> getSuggestions({
     required String text,
     required BoundingBox boundingBox,
     required SuggestOptions suggestOptions
-  }) {
-    final params = <String, dynamic>{
-      'sessionId': _nextSessionId++,
-      'text': text,
-      'boundingBox': boundingBox.toJson(),
-      'suggestOptions': suggestOptions.toJson(),
-    };
-    final result = _channel
-      .invokeMethod('getSuggestions', params)
-      .then((result) => SuggestSessionResult._fromJson(result));
+  }) async {
+    final session = await _initSession();
 
-    return SuggestResultWithSession._(
-      session: SuggestSession._(id: params['sessionId']),
-      result: result
-    );
+    return (session, session._getSuggestions(text: text, boundingBox: boundingBox, suggestOptions: suggestOptions));
+  }
+
+  /// Initialize session on native side for further use
+  static Future<SuggestSession> _initSession() async {
+    final id = _nextId++;
+
+    await _channel.invokeMethod('initSession', { 'id': id });
+
+    return SuggestSession._(id: id);
   }
 }
