+++ pubspec.yaml
@@ -4,18 +4,18 @@ version: 3.4.0
 homepage: https://github.com/Unact/yandex_mapkit
 
 environment:
-  sdk: ">=2.12.0 <4.0.0"
-  flutter: ">=3.0.0"
+  sdk: ">=3.0.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter_plugin_android_lifecycle: ^2.0.15
   equatable: ^2.0.5
-  collection: ^1.17.1
+  collection: ^1.18.0
   flutter:
     sdk: flutter
 
 dev_dependencies:
-  flutter_lints: ^2.0.2
+  flutter_lints: ^3.0.0
 
 flutter:
   plugin:
